+++ src/App/Fossa/PreflightChecks.hs
@@ -24,7 +24,7 @@ import Data.Text.IO qualified as TIO
 import Diag.Diagnostic (ToDiagnostic (..))
 import Effect.Logger (renderIt)
 import Errata (Errata (..))
-import Fossa.API.Types (ApiOpts, CustomBuildUploadPermissions (..), Organization (..), ProjectPermissionStatus (..), ReleaseGroupPermissionStatus (..), Subscription (..), TokenType (..), TokenTypeResponse (..))
+import Fossa.API.Types (ApiOpts, CustomBuildUploadPermissions (..), Organization (..), ProjectPermissionStatus (..), ReleaseGroupPermissionStatus (..), TokenType (..), TokenTypeResponse (..))
 import Path (
   File,
   Path,
@@ -78,7 +78,6 @@ preflightChecks cmd = context "preflight-checks" $ do
           ReportChecks -> do
             tokenType <- getTokenType
             fullAccessTokenCheck tokenType
-            premiumSubscriptionCheck org
           _ -> pure ()
   pure org
 
@@ -113,7 +112,7 @@ uploadBuildPermissionsCheck CustomBuildUploadPermissions{..} =
             $ fatal CreateTeamProjectPermissionErr
         InvalidCreateProjectOnlyToTeamPermission ->
           errDoc fossaConfigDocsUrl
-            . errHelp ("Ensure that you have specified a team to add this project to" :: Text)
+            . errHelp ("Please specify your team name, either in .fossa.yml or using the --team option with fossa analyze." :: Text)
             $ fatal CreateProjectOnlyToTeamPermissionErr
   where
     permissionHelpMsg :: Text
@@ -126,14 +125,6 @@ fullAccessTokenCheck TokenTypeResponse{..} = case tokenType of
       . errDoc apiTokenDocsUrl
       . errCtx ("You are currently using a `Push Only` API token" :: Text)
       $ fatal TokenTypeErr
-  _ -> pure ()
-
-premiumSubscriptionCheck :: Has Diagnostics sig m => Organization -> m ()
-premiumSubscriptionCheck Organization{..} = case orgSubscription of
-  Free ->
-    errHelp ("To proceed, please upgrade your subscription" :: Text)
-      . errCtx ("You currently have a free subscription" :: Text)
-      $ fatal SubscriptionTypeErr
   _ -> pure ()
 
 preflightCheckFileName :: Path Rel File
@@ -151,12 +142,6 @@ instance ToDiagnostic TokenTypeErr where
   renderDiagnostic TokenTypeErr =
     Errata (Just "Invalid API token type") [] $ Just "The action you are trying to perform requires a `Full Access` API token"
 
-data SubscriptionTypeErr = SubscriptionTypeErr
-instance ToDiagnostic SubscriptionTypeErr where
-  renderDiagnostic :: SubscriptionTypeErr -> Errata
-  renderDiagnostic SubscriptionTypeErr =
-    Errata (Just "Invalid subscription type") [] $ Just "The action you are trying to perform requires a premium subscription"
-
 projectPermissionErrHeader :: Text
 projectPermissionErrHeader = "Invalid project permission"
 
@@ -169,13 +154,13 @@ data ProjectPermissionErr
 instance ToDiagnostic ProjectPermissionErr where
   renderDiagnostic :: ProjectPermissionErr -> Errata
   renderDiagnostic CreateProjectPermissionErr =
-    Errata (Just projectPermissionErrHeader) [] $ Just "You do not have permission to create projects for your Organization"
+    Errata (Just projectPermissionErrHeader) [] $ Just "You do not have permission to create projects for your Organization."
   renderDiagnostic EditProjectPermissionErr =
-    Errata (Just projectPermissionErrHeader) [] $ Just "You do not have permission to edit projects for your Organization"
+    Errata (Just projectPermissionErrHeader) [] $ Just "You do not have permission to edit projects for your Organization."
   renderDiagnostic CreateTeamProjectPermissionErr =
-    Errata (Just projectPermissionErrHeader) [] $ Just "You do not have permission to create projects for the specified team"
+    Errata (Just projectPermissionErrHeader) [] $ Just "You do not have permission to create projects for the specified team."
   renderDiagnostic CreateProjectOnlyToTeamPermissionErr =
-    Errata (Just projectPermissionErrHeader) [] $ Just "You only have permission to create projects for your team"
+    Errata (Just projectPermissionErrHeader) [] $ Just "You only have permission to create projects for your team(s)."
 
 releaseGroupPermissionErrHeader :: Text
 releaseGroupPermissionErrHeader = "Invalid release group permission"